chore: reorganize files to /lib
Reorganizes and renames files for clarity.
