+++ lib/imap.js
@@ -3,8 +3,8 @@ const { ImapFlow } = require('imapflow');
 const debug = require('debug')('clownfish:imap');
 const cron = require('@ima-worldhealth/cron-scheduler');
 const pino = require('pino')(); // eslint-disable-line
-const api = require('../api');
-const utils = require('../utils');
+const api = require('./api');
+const utils = require('./utils');
 const { setupSMTPTransport } = require('./mail');
 
 // make the imapflow logger silent
